+++ subscriptions/__init__.py
@@ -1,0 +1,3 @@
+from .service import SubscriptionError, SubscriptionService
+
+__all__ = ["SubscriptionError", "SubscriptionService"]